OGR provider: moved QgsCPLErrorHandler to its own header file
This is being reused in gpkg data items
